V1.3- end point de atualização de coleira funcionando.    Refeita a forma da coleta de dados da "sitauação animal" para utilizar o AnimalModel.java.id_animal como chave e nao mais um valor chumbado.

+++ src/main/java/com/example/demo/Models/AnimalModel.java
@@ -19,7 +19,7 @@ import java.util.Set;
 public class AnimalModel implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_animal;
+    private long id_animal;
     private String nome;
     private Date idade;
     @Column(name = "coleira_id")
